+++ paddle_billing/Entities/Shared/AvailablePaymentMethods.py
@@ -2,10 +2,13 @@ from paddle_billing.PaddleStrEnum import PaddleStrEnum, PaddleStrEnumMeta
 
 
 class AvailablePaymentMethods(PaddleStrEnum, metaclass=PaddleStrEnumMeta):
-    Alipay: "AvailablePaymentMethods"     = 'alipay'
-    ApplePay: "AvailablePaymentMethods"   = 'apple_pay'
-    Bancontact: "AvailablePaymentMethods" = 'bancontact'
-    Card: "AvailablePaymentMethods"       = 'card'
-    GooglePay: "AvailablePaymentMethods"  = 'google_pay'
-    Ideal: "AvailablePaymentMethods"      = 'ideal'
-    Paypal: "AvailablePaymentMethods"     = 'paypal'
+    Alipay: "AvailablePaymentMethods"       = 'alipay'
+    ApplePay: "AvailablePaymentMethods"     = 'apple_pay'
+    Bancontact: "AvailablePaymentMethods"   = 'bancontact'
+    Card: "AvailablePaymentMethods"         = 'card'
+    GooglePay: "AvailablePaymentMethods"    = 'google_pay'
+    Ideal: "AvailablePaymentMethods"        = 'ideal'
+    Offline: "AvailablePaymentMethods"      = 'offline'
+    Paypal: "AvailablePaymentMethods"       = 'paypal'
+    Unknown: "AvailablePaymentMethods"      = 'unknown'
+    WireTransfer: "AvailablePaymentMethods" = 'wire_transfer'